Move file-watcher-store to file-system/controllers
