+++ src/file-system/controllers/store.ts
@@ -9,7 +9,6 @@ import { useFileTreeStore } from "../../file-explorer/controllers/file-tree-stor
 // This is an acceptable Zustand pattern, though it creates coupling between stores.
 // See: https://github.com/pmndrs/zustand/discussions/1319
 import { useBufferStore } from "../../stores/buffer-store";
-import { useFileWatcherStore } from "../../stores/file-watcher-store";
 import { useGitStore } from "../../stores/git-store";
 import { useProjectStore } from "../../stores/project-store";
 import { useRecentFoldersStore } from "../../stores/recent-folders-store";
@@ -38,6 +37,7 @@ import { getGitStatus } from "../../utils/git";
 import { isDiffFile, parseRawDiffContent } from "../../utils/git-diff-parser";
 import type { FileEntry } from "../models/app";
 import type { FsActions, FsState } from "../models/interface";
+import { useFileWatcherStore } from "./file-watcher-store";
 import { openFolder, readDirectory } from "./platform";
 import { shouldIgnore, updateDirectoryContents } from "./utils";
